Fix possible bug introduced in r88: warning: comparison of unsigned expression < 0 is always false
git-svn-id: https://svn.il.fontys.nl/usersvn/fishtank/svn/trunk@73 bb1b31de-baa9-489e-af1c-03370f2903b8

+++ Vissen Project Aquarium/imagereceiver.cpp
@@ -304,7 +304,7 @@ bool ImageReceiver::ReceiveSegment(){
 
 	const size_t wanted_buffer_size = image_size>header_size ? image_size : header_size;
 	if(buffer.size()<wanted_buffer_size)buffer.resize(wanted_buffer_size);
-	const size_t bytes_to_receive=wanted_buffer_size-buffered_bytes;
+	const int bytes_to_receive=wanted_buffer_size-buffered_bytes;
 	if(bytes_to_receive<0){
 		std::cerr<<"Stream is corrupt(2), closing connection!"<<std::endl;
 		m_socket_stream.Close();
